Create Edge sum (Nallas)
A set of numbers forming a matrix N*N is passes as input. The program has to print the sum of
numbers which are not along the edges.
Input Format:
The first line will contain the value of N
The next lines will contain N numbers each separated by one or more spaces
Boundary Conditions:
3 <= N <= 10
Value of a given number in the matrix is from 0 to 99999999
Output Format:
The sum of the numbers which are not along the edges.
Example Input / Output 1:
Input:
3
5 6 7
8 9 1
2 3 4
Output:
9
Explanation:
All numbers except 9 are along the edges. Hence the sum is 9 which is printed as output.